% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadPartialCorProt}
\alias{loadPartialCorProt}
\title{Load PartialCorProt data}
\usage{
loadPartialCorProt()
}
\value{
A data.frame
}
\description{
This function loads the PartialCorProt dataset from ExperimentHub.
}
\examples{
proteins <- loadPartialCorProt()
if (!inherits(proteins, "try-error")) head(proteins)
}
