% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadPartialCorGenes}
\alias{loadPartialCorGenes}
\title{Load PartialCorGenes data}
\usage{
loadPartialCorGenes()
}
\value{
A data.frame
}
\description{
This function loads the PartialCorGenes dataset from ExperimentHub.
}
\examples{
genes <- loadPartialCorGenes()
if (!inherits(genes, "try-error")) head(genes)
}
